Improve API Reference in Tech Radar Plugin
