+++ plugins/tech-radar/api-report.md
@@ -9,138 +9,114 @@ import { ApiRef } from '@backstage/core-plugin-api';
 import { BackstagePlugin } from '@backstage/core-plugin-api';
 import { RouteRef } from '@backstage/core-plugin-api';
 
+// Warning: (ae-missing-release-tag) "MovedState" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public
+export enum MovedState {
+  Down = -1,
+  NoChange = 0,
+  Up = 1,
+}
+
 // Warning: (ae-missing-release-tag) "RadarEntry" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
 //
-// @public (undocumented)
+// @public
 export interface RadarEntry {
-  // (undocumented)
   description?: string;
-  // (undocumented)
   id: string;
-  // (undocumented)
   key: string;
-  // (undocumented)
   quadrant: string;
-  // (undocumented)
   timeline: Array<RadarEntrySnapshot>;
-  // (undocumented)
   title: string;
-  // (undocumented)
   url: string;
 }
 
 // Warning: (ae-missing-release-tag) "RadarEntrySnapshot" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
 //
-// @public (undocumented)
+// @public
 export interface RadarEntrySnapshot {
-  // (undocumented)
   date: Date;
-  // (undocumented)
   description?: string;
-  // Warning: (ae-forgotten-export) The symbol "MovedState" needs to be exported by the entry point index.d.ts
-  //
-  // (undocumented)
   moved?: MovedState;
-  // (undocumented)
   ringId: string;
 }
 
 // Warning: (ae-missing-release-tag) "RadarQuadrant" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
 //
-// @public (undocumented)
+// @public
 export interface RadarQuadrant {
-  // (undocumented)
   id: string;
-  // (undocumented)
   name: string;
 }
 
 // Warning: (ae-missing-release-tag) "RadarRing" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
 //
-// @public (undocumented)
+// @public
 export interface RadarRing {
-  // (undocumented)
   color: string;
-  // (undocumented)
   id: string;
-  // (undocumented)
   name: string;
 }
 
 // Warning: (ae-missing-release-tag) "RadarPage" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
 //
-// @public (undocumented)
-export const Router: {
-  ({ title, subtitle, pageTitle, ...props }: TechRadarPageProps): JSX.Element;
-  defaultProps: {
-    title: string;
-    subtitle: string;
-    pageTitle: string;
-  };
-};
+// @public
+export function Router(props: TechRadarPageProps): JSX.Element;
 
 // Warning: (ae-missing-release-tag) "TechRadarApi" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
 //
-// @public (undocumented)
+// @public
 export interface TechRadarApi {
   load: (id: string | undefined) => Promise<TechRadarLoaderResponse>;
 }
 
 // Warning: (ae-missing-release-tag) "techRadarApiRef" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
 //
-// @public (undocumented)
+// @public
 export const techRadarApiRef: ApiRef<TechRadarApi>;
 
 // Warning: (ae-missing-release-tag) "RadarComponent" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
 //
-// @public (undocumented)
-export const TechRadarComponent: (
-  props: TechRadarComponentProps,
-) => JSX.Element;
+// @public
+export function TechRadarComponent(props: TechRadarComponentProps): JSX.Element;
 
 // Warning: (ae-missing-release-tag) "TechRadarComponentProps" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
 //
-// @public (undocumented)
+// @public
 export interface TechRadarComponentProps {
-  // (undocumented)
   height: number;
-  // (undocumented)
   id?: string;
-  // (undocumented)
   searchText?: string;
-  // (undocumented)
   svgProps?: object;
-  // (undocumented)
   width: number;
 }
 
 // Warning: (ae-missing-release-tag) "TechRadarLoaderResponse" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
 //
-// @public (undocumented)
+// @public
 export interface TechRadarLoaderResponse {
-  // (undocumented)
   entries: RadarEntry[];
-  // (undocumented)
   quadrants: RadarQuadrant[];
-  // (undocumented)
   rings: RadarRing[];
 }
 
 // Warning: (ae-missing-release-tag) "TechRadarPage" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
 //
-// @public (undocumented)
-export const TechRadarPage: {
-  ({ title, subtitle, pageTitle, ...props }: TechRadarPageProps): JSX.Element;
-  defaultProps: {
-    title: string;
-    subtitle: string;
-    pageTitle: string;
-  };
+// @public
+export const TechRadarPage: Router;
+
+// Warning: (ae-missing-release-tag) "TechRadarPageProps" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public
+export type TechRadarPageProps = TechRadarComponentProps & {
+  title?: string;
+  subtitle?: string;
+  pageTitle?: string;
 };
 
 // Warning: (ae-missing-release-tag) "techRadarPlugin" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
 //
-// @public (undocumented)
+// @public
 const techRadarPlugin: BackstagePlugin<
   {
     root: RouteRef<undefined>;
@@ -149,8 +125,4 @@ const techRadarPlugin: BackstagePlugin<
 >;
 export { techRadarPlugin as plugin };
 export { techRadarPlugin };
-
-// Warnings were encountered during analysis:
-//
-// src/components/RadarPage.d.ts:9:5 - (ae-forgotten-export) The symbol "TechRadarPageProps" needs to be exported by the entry point index.d.ts
 ```
