+++ plugins/tech-radar/src/components/RadarPage.tsx
@@ -23,7 +23,7 @@ import {
 } from '@backstage/core-components';
 import { Grid, Input, makeStyles } from '@material-ui/core';
 import React from 'react';
-import { TechRadarComponentProps } from '../api';
+import { TechRadarPageProps } from '../api';
 import RadarComponent from '../components/RadarComponent';
 
 const useStyles = makeStyles(() => ({
@@ -32,18 +32,16 @@ const useStyles = makeStyles(() => ({
   },
 }));
 
-export type TechRadarPageProps = TechRadarComponentProps & {
-  title?: string;
-  subtitle?: string;
-  pageTitle?: string;
-};
-
-export const RadarPage = ({
-  title,
-  subtitle,
-  pageTitle,
-  ...props
-}: TechRadarPageProps): JSX.Element => {
+/**
+ * Main Page of Tech Radar
+ */
+export function RadarPage(props: TechRadarPageProps) {
+  const {
+    title = 'Tech Radar',
+    subtitle = 'Pick the recommended technologies for your projects',
+    pageTitle = 'Company Radar',
+    ...componentProps
+  } = props;
   const classes = useStyles();
   const [searchText, setSearchText] = React.useState('');
 
@@ -66,16 +64,10 @@ export const RadarPage = ({
         </ContentHeader>
         <Grid container spacing={3} direction="row">
           <Grid item xs={12} sm={6} md={4}>
-            <RadarComponent searchText={searchText} {...props} />
+            <RadarComponent searchText={searchText} {...componentProps} />
           </Grid>
         </Grid>
       </Content>
     </Page>
   );
-};
-
-RadarPage.defaultProps = {
-  title: 'Tech Radar',
-  subtitle: 'Pick the recommended technologies for your projects',
-  pageTitle: 'Company Radar',
-};
+}
