+++ plugins/tech-radar/src/plugin.ts
@@ -28,6 +28,9 @@ const rootRouteRef = createRouteRef({
   title: 'Tech Radar',
 });
 
+/**
+ * Tech Radar plugin instance
+ */
 export const techRadarPlugin = createPlugin({
   id: 'tech-radar',
   routes: {
@@ -36,6 +39,13 @@ export const techRadarPlugin = createPlugin({
   apis: [createApiFactory(techRadarApiRef, new SampleTechRadarApi())],
 });
 
+/**
+ * Main Tech Radar Page
+ *
+ * @remarks
+ *
+ * Uses {@link TechRadarPageProps} as props
+ */
 export const TechRadarPage = techRadarPlugin.provide(
   createRoutableExtension({
     component: () => import('./components/RadarPage').then(m => m.RadarPage),
